requirements.txt
google-genai==1.12.1
flask==3.0.3
python-dotenv==1.1.0
python-telegram-bot==22.6
httpx>=0.27,<0.29
